requirements.txt
torch
torchvision
Pillow
requests
safetensors
tqdm
open_clip_torch
accelerate
transformers>=4.27.1